Refactor: Change the file structure.
Separate othero/core.py as follows.
    othero/core/board.py: count_soss, get_positions_in_sos
    othero/core/pos.py: advance_pos, is_in_board
    othero/core/sog.py: duplicate_sog, sog_to_string, string_to_sog
    othero/core/types.py: Direction, SOS, Disk

    othero/rule/forward.py: is_sos_change_valid,
calc_sog_after_sos_changed
    othero/rule/utils.py: calc_sid, calc_all_sids
