+++ othero/game/utils.py
@@ -5,14 +5,15 @@
 # This source code is licensed under the MIT License found in
 # the LICENSE file in the root directory of this source tree.
 
-from othero import core
+from othero.core import board, sog, types
+from othero.rule import forward
 from othero.display import utils
 
 INIT_SOG = [ \
-    [core.SOS.BLANK , core.SOS.BLANK, core.SOS.BLANK, core.SOS.BLANK], \
-    [core.SOS.BLANK , core.SOS.DARK , core.SOS.LIGHT, core.SOS.BLANK], \
-    [core.SOS.BLANK , core.SOS.LIGHT, core.SOS.DARK , core.SOS.BLANK], \
-    [core.SOS.BLANK , core.SOS.BLANK, core.SOS.BLANK, core.SOS.BLANK] \
+    [types.SOS.BLANK , types.SOS.BLANK, types.SOS.BLANK, types.SOS.BLANK], \
+    [types.SOS.BLANK , types.SOS.DARK , types.SOS.LIGHT, types.SOS.BLANK], \
+    [types.SOS.BLANK , types.SOS.LIGHT, types.SOS.DARK , types.SOS.BLANK], \
+    [types.SOS.BLANK , types.SOS.BLANK, types.SOS.BLANK, types.SOS.BLANK] \
 ]
 
 class InvalidDiskColorError(Exception):
@@ -20,7 +21,7 @@ class InvalidDiskColorError(Exception):
         self.disk = disk
     
     def __str__(self):
-        sdisk = utils.visualize_sos(core.Disk.toSOS(self.disk))
+        sdisk = utils.visualize_sos(types.Disk.toSOS(self.disk))
         return f"disk: {sdisk}"
 
 class InvalidDiskPositionError(Exception):
@@ -36,7 +37,7 @@ class InvalidDiskPositionError(Exception):
     
     def __str__(self):
         ssog = utils.visualize_sog(self.sog)
-        sdisk = utils.visualize_sos(core.Disk.toSOS(self.disk))
+        sdisk = utils.visualize_sos(types.Disk.toSOS(self.disk))
         return f"""
         sog:  {ssog}
         pos:  {self.pos}
@@ -51,7 +52,7 @@ class Game():
         self.sog = create_sog()
         self.dark_player_name = dark_player_name
         self.light_player_name = light_player_name
-        self.next_disk = core.Disk.DARK
+        self.next_disk = types.Disk.DARK
     
     def put(self, pos, disk):
         if disk != self.next_disk:
@@ -59,17 +60,17 @@ class Game():
         self.sog = calc_sog_after_put_disk(self.sog, pos, disk)
     
     def turn(self):
-        if self.next_disk == core.Disk.DARK:
-            self.next_disk = core.Disk.LIGHT
+        if self.next_disk == types.Disk.DARK:
+            self.next_disk = types.Disk.LIGHT
         else:
-            self.next_disk = core.Disk.DARK
+            self.next_disk = types.Disk.DARK
     
     def isAbleToPut(self):
         avail_poss = get_available_positions(self.sog, self.next_disk)
         return len(avail_poss) != 0
 
     def countDisks(self):
-        ndark, nlight, _ = core.count_soss(self.sog)
+        ndark, nlight, _ = board.count_soss(self.sog)
         return ndark, nlight
 
 def create_sog():
@@ -80,7 +81,7 @@ def create_sog():
         [[othero.core.SOS]]:
             The newly created sog in the initial state.
     """
-    return core.duplicate_sog(INIT_SOG)
+    return sog.duplicate_sog(INIT_SOG)
 
 def is_put_disk_valid(sog, pos, disk):
     """
@@ -104,7 +105,7 @@ def is_put_disk_valid(sog, pos, disk):
         bool:
             The validity of putting <disk>.
     """
-    return core.is_sos_change_valid(sog, pos, core.Disk.toSOS(disk))
+    return forward.is_sos_change_valid(sog, pos, types.Disk.toSOS(disk))
 
 def calc_sog_after_put_disk(sog, pos, disk):
     """
@@ -137,7 +138,7 @@ def calc_sog_after_put_disk(sog, pos, disk):
     if not is_put_disk_valid(sog, pos, disk):
         raise InvalidDiskPositionError(sog, pos, disk)
 
-    return core.calc_sog_after_sos_changed(sog, pos, core.Disk.toSOS(disk))
+    return forward.calc_sog_after_sos_changed(sog, pos, types.Disk.toSOS(disk))
 
 def get_available_positions(sog, disk):
     """
@@ -154,8 +155,8 @@ def get_available_positions(sog, disk):
         [(int, int)]:
             List of positions in <sog> where putting <disk> is allowed.
     """
-    sos = core.Disk.toSOS(disk)
+    sos = types.Disk.toSOS(disk)
 
-    blank_poss = core.get_positions_in_sos(sog, core.SOS.BLANK)
+    blank_poss = board.get_positions_in_sos(sog, types.SOS.BLANK)
     return [pos for pos in blank_poss \
-                if core.is_sos_change_valid(sog, pos, sos)]
+                if forward.is_sos_change_valid(sog, pos, sos)]
